Prusa Carriage Mod
This is a mod to the ecksbot X-Carriage that will allow it to be printed
as a drop in replacement to the regular carriage found on numerous Prusa
Mendel variants (i.e. prusa 1, prusa it2, wallace, mix G1, etc.)  There
are 2 version, one with a printed hook to further reduce hardware
commits and one withou the hook so that different extruder blocks can be
used.